requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
email-validator>=2.1.0
python-dotenv>=1.0.0
huggingface-hub>=0.19.0